# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ExtractedPdfText {
  text: string;
  pageCount: number;
  fileName: string;
}

export interface GenerateStudyKitBody {
  text: string;
  fileName?: string;
}

export interface Flashcard {
  question: string;
  answer: string;
}

export interface QuizQuestion {
  question: string;
  answer: string;
}

export interface StudyPlanDay {
  day: number;
  title: string;
  tasks: string[];
}

export interface StudyKit {
  summary: string;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
  studyPlan: StudyPlanDay[];
  /** @nullable */
  fileName?: string | null;
}

export type UploadPdfBody = {
  file: Blob;
};