# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate AI study kit from PDF text
 */
export const GenerateStudyKitBody = zod.object({
  text: zod.string(),
  fileName: zod.string().optional(),
});

export const GenerateStudyKitResponse = zod.object({
  summary: zod.string(),
  flashcards: zod.array(
    zod.object({
      question: zod.string(),
      answer: zod.string(),
    }),
  ),
  quizQuestions: zod.array(
    zod.object({
      question: zod.string(),
      answer: zod.string(),
    }),
  ),
  studyPlan: zod.array(
    zod.object({
      day: zod.number(),
      title: zod.string(),
      tasks: zod.array(zod.string()),
    }),
  ),
  fileName: zod.string().nullish(),
});

/**
 * @summary Upload and extract text from a PDF
 */
export const UploadPdfBody = zod.object({
  file: zod.instanceof(File),
});

export const UploadPdfResponse = zod.object({
  text: zod.string(),
  pageCount: zod.number(),
  fileName: zod.string(),
});