# lib/api-zod/src/generated/types/studyKit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from "./flashcard";
import type { QuizQuestion } from "./quizQuestion";
import type { StudyPlanDay } from "./studyPlanDay";

export interface StudyKit {
  summary: string;
  flashcards: Flashcard[];
  quizQuestions: QuizQuestion[];
  studyPlan: StudyPlanDay[];
  /** @nullable */
  fileName?: string | null;
}